fix(GraphQL): Organization/Reseller will see all Customer `Location`s. (#857)
`OwnedByOrganization` was added to `Location` in 161c73b43216f56711f6fad4d3648138a1785bd4, but seems accedentaly, to fix new tests. Right now there are no reasons to check organization for `Location` model, because Location(s) are visible only as nested data. Moreover, `LocationsProcessor` (Recalculator) doesn't add `ResellerLocations` for customers of the reseller (and never did) - this is the reason of the issue.

Closes: #856